==> 01-Intro.Rmd <==
# Introduction {-}

In the present study, we assessed genotypic diversity within closely-related sympatric tree species belonging to the widespread tropical tree species complex *Symphonia globulifera*.
We addressed the fine-scale spatial and temporal genetic adaptations of individuals through differential growth strategies in response to forest gap dynamics. 
We finally compared the breadth of successional niches encountered by *Symphonia* species to other locally abundant species. 
Combining tree diameter censuses, indirect measures of light environment of the recent past and present, and single nucleotide polymorphisms (SNPs), 
we used population genomics, environmental association analyses, genome wide association and growth modelling to address the following questions:

* Are individual genotypes structured by the mosaic of light and competition environments resulting from forest gap dynamics?
* Is the growth of individuals determined by genotypes?
* Is there an association between genotypic adaptations to gap dynamics and to growth?
* How are genotypic adaptations to gap dynamics and to growth structured in  time, i.e., across life stages?
* Are breadths of successional niches for *Symphonia* species wider than those of other locally abundant species?
